grpc-ui/gui/requirements.txt
flask ~= 1.1.2
grpcio-tools ~= 1.30
Jinja2 ~= 2.11
pytest ~= 5.4